pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "factoryverse"
version = "0.6.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiodocker>=0.24.0",
    "duckdb>=1.4.3",
    "factorio-rcon-py>=2.1.3",
    "nbformat>=5.10.4",
    "numpy>=2.3.5",
    "openai>=2.9.0",
    "pandas>=2.3.3",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "watchdog>=6.0.0",
    "websocket-client>=1.9.0",
]

[project.scripts]
factoryverse = "FactoryVerse.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/FactoryVerse"]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "shapely>=2.1.2",
    "tiktoken>=0.12.0",
]
